+++ src/app/(front-end)/page.tsx
@@ -1,13 +1,13 @@
 import { LogOutButton } from '@/auth/nextjs/components/logout-button';
 import { getCurrentUser } from '@/auth/nextjs/current-user';
-import { Button } from '@/components/ui/button';
 import {
+  Button,
   Card,
   CardDescription,
   CardFooter,
   CardHeader,
   CardTitle,
-} from '@/components/ui/card';
+} from '@/components';
 import Link from 'next/link';
 
 export default async function HomePage() {
