+++ src/auth/nextjs/components/logout-button.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components';
 import { logOut } from '../actions';
 
 export function LogOutButton() {
