+++ src/components/index.ts
@@ -1,13 +1,1 @@
-export * from './ui/button';
-export * from './ui/card';
-export * from './ui/form';
-export * from './ui/form-status';
-export * from './ui/form-info';
-export * from './ui/input';
-export * from './ui/label';
-export * from './ui/password-input';
-export * from './ui/form-header';
-export * from './ui/icon-badge';
-export * from './ui/container';
-export * from './ui/section';
-export * from './ui/separator';
+export * from './ui';
